(PE-11531) Always manage agent, even if already upgraded
Prior to this commit, if a pe agent was already up to date,
we would not attempt to manage any of the resources for it.
This commit removes the early abort if a PE agent is up to date,
and will leave any PE-specific managed resources that need to not
be managed.